[FIX] web: show warning before closing view form
In a view form, an onchange could only be triggered just before closing
the modal. Hence people may not be made aware of a warning which may
have informed or made them change what they were about to do.

This change cancel a save when a warning is happening.

closes #7089
opw-642067